refactor lumino-1.py and lumino-2.py into lumino.py, which is parametrized
